% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genexpMatrix_example}
\alias{genexpMatrix_example}
\title{Matrix with log-transformed cpm values of 500 references genes identified in TCGA studies.}
\format{
A matrix with 500 rows and 9,724 columns.
}
\usage{
genexpMatrix_example
}
\description{
A dataset containing gene expression data of the 9,724 tumor samples
accross the 32 different TCGA tumor studies.
}
\keyword{datasets}
