% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_intra_project_sigma}
\alias{mean_intra_project_sigma}
\title{Dataframe with the estimated mean 'within study' SD across the 32 included tumor studies.}
\format{

}
\usage{
mean_intra_project_sigma
}
\description{
Dataframe with the estimated mean 'within study' SD across the 32 included tumor studies.
}
\keyword{datasets}
