% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleMeta_example}
\alias{sampleMeta_example}
\title{Contains the abbreviated sample meta information.}
\format{
A data frame with 9,724 rows and 2 variables:
\describe{
  \item{Study_ID}{Unique ID of the study}
  \item{Sample_ID}{Unique ID of the sample, must exist in corresponding genexpression data}
}
}
\source{
\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}
}
\usage{
sampleMeta_example
}
\description{
This data describes which TCGA study each sample belongs to.
}
\keyword{datasets}
